feat: add the score screen to the game
When the travel is finished, the game will show the score screen with the
final score over 100 points.

Change a little bit of the structure of the code to dissable the game
screen when the score screen is shown, but with the capability to
restart the game or change the level.